refactor: structure server source
