+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "feedback-fusion"
 version = "0.5.0"
-edition = "2021"
+edition = "2024"
 license = "MIT"
 
 [dependencies]
@@ -21,6 +21,7 @@ cached = { version = "0.53.0", features = ["async", "proc_macro"] }
 chrono = { version = "0.4.38", features = ["serde"] }
 dashmap = { version = "6.1.0", features = ["rayon"] }
 derivative = "2.2.0"
+educe = "0.6.0"
 envy = "0.4.2"
 futures = "0.3.30"
 getset = "0.1.2"
