+++ src/authorization/oidc/claims.rs
@@ -1,9 +1,9 @@
-//SPDX-FileCopyrightText: 2023 OneLiteFeatherNet
+//SPDX-FileCopyrightText: 2025 OneLiteFeatherNet
 //SPDX-License-Identifier: MIT
 
 //MIT License
 
-// Copyright (c) 2023 OneLiteFeatherNet
+// Copyright (c) 2025 OneLiteFeatherNet
 
 //Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
 //associated documentation files (the "Software"), to deal in the Software without restriction,
@@ -20,95 +20,16 @@
 //DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-use std::str::FromStr;
-
 use crate::prelude::*;
 
-use aliri::{
-    jwa,
-    jwt::{self, CoreValidator},
-};
+use aliri::jwt::{self};
 use aliri_clock::UnixTime;
-use aliri_oauth2::{Authority, HasScope, Scope};
-use aliri_tower::OnJwtError;
-use openidconnect::{
-    core::{CoreClient, CoreJwsSigningAlgorithm, CoreProviderMetadata},
-    ClientId, IssuerUrl,
-};
+use aliri_oauth2::{HasScope, Scope};
+
 use serde::{
     de::{MapAccess, Visitor},
     Deserializer,
 };
-use tokio::runtime::Handle;
-use tonic::{body::BoxBody, Status};
-
-pub async fn authority() -> Result<(Authority, CoreClient)> {
-    // sadly aliri does not support oidc yet, so we have to do the config stuff manually :(((((
-    // discover the oidc endpoints
-    let issuer = IssuerUrl::new(CONFIG.oidc().provider().clone()).map_err(|error| {
-        FeedbackFusionError::ConfigurationError(format!("Invalid discovery url: {}", error))
-    })?;
-    let metadata = CoreProviderMetadata::discover_async(
-        issuer.clone(),
-        openidconnect::reqwest::async_http_client,
-    )
-    .await
-    .map_err(|error| {
-        FeedbackFusionError::ConfigurationError(format!("Invalid oidc endpoint: {}", error))
-    })?;
-    // extract the jwks
-    let jwks_url = metadata.jwks_uri().url();
-    // extract the algorithms
-    let algorithms = metadata
-        .id_token_signing_alg_values_supported()
-        .iter()
-        .filter_map(|key| match key {
-            CoreJwsSigningAlgorithm::HmacSha256 => Some(jwa::Algorithm::HS256),
-            CoreJwsSigningAlgorithm::HmacSha384 => Some(jwa::Algorithm::HS384),
-            CoreJwsSigningAlgorithm::HmacSha512 => Some(jwa::Algorithm::HS512),
-            CoreJwsSigningAlgorithm::EcdsaP256Sha256 => Some(jwa::Algorithm::PS256),
-            CoreJwsSigningAlgorithm::EcdsaP384Sha384 => Some(jwa::Algorithm::PS384),
-            CoreJwsSigningAlgorithm::EcdsaP521Sha512 => Some(jwa::Algorithm::PS512),
-            CoreJwsSigningAlgorithm::RsaSsaPkcs1V15Sha256 => Some(jwa::Algorithm::RS256),
-            CoreJwsSigningAlgorithm::RsaSsaPkcs1V15Sha384 => Some(jwa::Algorithm::RS384),
-            CoreJwsSigningAlgorithm::RsaSsaPkcs1V15Sha512 => Some(jwa::Algorithm::RS512),
-            _ => None,
-        })
-        .collect::<Vec<jwa::Algorithm>>();
-
-    // build the validator
-    let mut validator = CoreValidator::default()
-        .add_allowed_audience(
-            jwt::Audience::from_str(CONFIG.oidc().audience().as_str())
-                .expect("Invalid oidc audience"),
-        )
-        .require_issuer(
-            jwt::Issuer::from_str(
-                CONFIG
-                    .oidc()
-                    .issuer()
-                    .clone()
-                    .unwrap_or(issuer.clone().to_string())
-                    .as_str(),
-            )
-            .unwrap(),
-        );
-    for algorithm in algorithms {
-        validator = validator.add_approved_algorithm(algorithm);
-    }
-
-    // build the authority
-    let authority = Authority::new_from_url(jwks_url.to_string(), validator)
-        .await
-        .unwrap();
-
-    // we do not use any routes requireing client authentication therefore we can just use dummys
-    // here
-    let client =
-        CoreClient::from_provider_metadata(metadata, ClientId::new("dummy".to_owned()), None);
-
-    Ok((authority, client))
-}
 
 // we do not know the group claim names during the compile time, therefore we do have to use this
 // huge custom deserializer
@@ -249,77 +170,5 @@ impl HasScope for OIDCClaims {
 impl OIDCClaims {
     pub fn groups(&self) -> &Vec<String> {
         &self.groups
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-pub struct AuthorizedService<S>(pub S);
-
-impl<S, T> tonic::server::NamedService
-    for AuthorizedService<tower_http::validate_request::ValidateRequestHeader<S, T>>
-where
-    S: tonic::server::NamedService,
-{
-    const NAME: &'static str = S::NAME;
-}
-
-impl<S, R> tower::Service<R> for AuthorizedService<S>
-where
-    S: tower::Service<R>,
-{
-    type Error = S::Error;
-    type Future = S::Future;
-    type Response = S::Response;
-
-    #[inline]
-    fn poll_ready(
-        &mut self,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<std::result::Result<(), Self::Error>> {
-        self.0.poll_ready(cx)
-    }
-
-    #[inline]
-    fn call(&mut self, req: R) -> Self::Future {
-        self.0.call(req)
-    }
-}
-
-#[derive(Clone)]
-pub struct OIDCErrorHandler {
-    authority: Authority,
-}
-
-impl From<Authority> for OIDCErrorHandler {
-    fn from(authority: Authority) -> Self {
-        Self { authority }
-    }
-}
-
-impl OnJwtError for OIDCErrorHandler {
-    type Body = BoxBody;
-
-    fn on_jwt_invalid(
-        &self,
-        _error: aliri::error::JwtVerifyError,
-    ) -> opentelemetry_http::Response<Self::Body> {
-        Status::unauthenticated("unauthenticated").into_http()
-    }
-
-    fn on_no_matching_jwk(&self) -> opentelemetry_http::Response<Self::Body> {
-        warn!("No matching jwk for request found, refreshing jwks...");
-
-        let handle = Handle::current();
-        let authority = self.authority.clone();
-
-        // theoretically aliri does spawn a thread itself but as it requires a async runtime
-        // context we do have to spawn a new runtim thread ourselves
-        handle.spawn(async move { authority.refresh().await.ok() });
-
-        Status::unauthenticated("unauthenticated").into_http()
-    }
-
-    fn on_missing_or_malformed(&self) -> opentelemetry_http::Response<Self::Body> {
-        Status::unauthenticated("unauthenticated").into_http()
     }
 }
